+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This changelog suppose to follow rules defined in the [changelog.md](https://cha
 
 ## 0.2.0 - TBD
 
+- **Changed**: VOG1 BLE protocol introduced
+- **Changed**: WiFI / BLE refactor
+
 ## 0.1.0 - 2024-06-10
 
 Initial release with cute structure for simple MQTT publishing of measurements from BLE advertisements from IoT sensors.